requirements.txt
streamlit
pandas
numpy
rapidfuzz
requests
html2text
tqdm
urllib3
beautifulsoup4
langchain==0.3.23
langchain_openai==0.3.9
langchain_community
langgraph
tavily-python
python-dotenv
pymupdf
matplotlib
networkx
altair
seaborn
openai